Don't double-stringify the branch list
ParseRcs was turning the branches list into a string before calling
setBranches; setBranches then treated that string as a list and
re-stringified it.  The result was that a branch list that contained
1.1.1.1 and 1.2.1.1 would become '1 . 1 . 1 . 1   1 . 2 . 1 . 1'
rather than the correct '1.1.1.1 1.2.1.1'.

This is technically a functionality change, but presumably there is no
existing code that depends on the incorrect behavior.